requirements.txt
sqlmodel
fastapi
uvicorn
pydantic
requests
sqlalchemy
psycopg2
python-dotenv
typing